# Optimize prefill speed for Qwen3-MOE GGUF (ISQ) & adjustable prefill chunk size