This PR updates the underlying Candle, where we optimized the previous indexed_moe_forward kernel to speed up the prefill stage in qwen3 moe models. Relevant code can be found here: https://github.com/guoqingbao/candle/commit/cf5b1ad79558ac2de208426a6c0f80b1ee29dd5d

Additionally, this PR makes the prefill chunk size configurable.

**Testing results:**

```
--- Performance Metrics ---
⏱️ Prompt tokens: 24328 in 23.32s (1043.05 tokens/s)
⏱️ Decoded tokens: 2608 in 49.34s (52.86 tokens/s)
```

Prefill performance improved from ~600 tokens/s to over 1000 tokens/s on two A100s when running Qwen3-MoE 32B (ISQ Q4K).

Related #253 @sempervictus